requirements.txt
pynput==1.7.7
pillow==10.3.0
pyautogui==0.9.54
